requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.5
dnspython==2.4.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
Flask==2.2.3
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fpdf==1.7.2
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
pdfkit==1.0.0
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0
reportlab==4.0.4
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.18
typing-extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.1
